add color conversion to generate ng_link
