+++ code/aind_smartspim_stitch/utils/utils.py
@@ -513,3 +513,28 @@ def generate_timestamp(time_format: str = "%Y-%m-%d_%H-%M-%S") -> str:
         moment in string format.
     """
     return datetime.now().strftime(time_format)
+
+def wavelength_to_hex(wavelength: int) -> int:
+    # Each wavelength key is the upper bound to a wavelgnth band. 
+    # Wavelengths range from 380-750nm.
+    # Color map wavelength/hex pairs are generated by sampling along a CIE diagram arc.
+    color_map = {460: 0x690afe,  # Purple
+                 470: 0x3f2efe,  # Blue-Purple
+                 480: 0x4b90fe,  # Blue
+                 490: 0x59d5f8,  # Blue-Green
+                 500: 0x5df8d6,  # Green
+                 520: 0x5afeb8,  # Green
+                 540: 0x58fea1,  # Green
+                 560: 0x51ff1e,  # Green
+                 565: 0xbbfb01,  # Green-Yellow
+                 575: 0xe9ec02,  # Yellow
+                 580: 0xf5c503,  # Yellow-Orange
+                 590: 0xf39107,  # Orange
+                 600: 0xf15211,  # Orange-Red
+                 620: 0xf0121e,  # Red
+                 750: 0xf00050}  # Pink
+
+    for ub, hex_val in color_map.items():
+        if wavelength < ub:     # Exclusive
+            return hex_val
+    return hex_val      # hex_val is set to the last color in for loop
