+++ code/aind_smartspim_stitch/zarr_converter/zarr_converter.py
@@ -25,12 +25,12 @@ from natsort import natsorted
 from numcodecs import blosc
 from skimage.io import imread as sk_imread
 
+import utils
 from .zarr_converter_params import ZarrConvertParams, get_default_config
 
 PathLike = Union[str, Path]
 ArrayLike = Union[dask.array.core.Array, np.ndarray]
 blosc.use_threads = False
-
 
 def add_leading_dim(data: ArrayLike) -> ArrayLike:
     """
@@ -610,31 +610,6 @@ def get_sample_img(directory_structure: dict) -> ArrayLike:
 
     return sample_img
 
-def wavelength_to_hex(wavelength: int) -> int:
-    # Each wavelength key is the upper bound to a wavelgnth band. 
-    # Wavelengths range from 380-750nm.
-    # Color map wavelength/hex pairs are generated by sampling along a CIE diagram arc.
-    color_map = {460: 0x690afe,  # Purple
-                 470: 0x3f2efe,  # Blue-Purple
-                 480: 0x4b90fe,  # Blue
-                 490: 0x59d5f8,  # Blue-Green
-                 500: 0x5df8d6,  # Green
-                 520: 0x5afeb8,  # Green
-                 540: 0x58fea1,  # Green
-                 560: 0x51ff1e,  # Green
-                 565: 0xbbfb01,  # Green-Yellow
-                 575: 0xe9ec02,  # Yellow
-                 580: 0xf5c503,  # Yellow-Orange
-                 590: 0xf39107,  # Orange
-                 600: 0xf15211,  # Orange-Red
-                 620: 0xf0121e,  # Red
-                 750: 0xf00050}  # Pink
-
-    for ub, hex_val in color_map.items():
-        if wavelength < ub:     # Exclusive
-            return hex_val
-    return hex_val      # hex_val is set to the last color in for loop
-
 class ZarrConverter:
     """
     Class to convert smartspim datasets to the zarr format
@@ -693,7 +668,7 @@ class ZarrConverter:
 
         for channel_str in self.channels:
             em_wav: int = int(channel_str.split('_')[-1])
-            em_hex: int = wavelength_to_hex(em_wav)
+            em_hex: int = utils.wavelength_to_hex(em_wav)
             self.channel_colors.append(em_hex)
 
         # get_blosc_codec(writer_config['codec'], writer_config['clevel'])
